+++ test/verify/test_verify.py
@@ -1,7 +1,7 @@
 from fastapi.testclient import TestClient
 from sqlmodel import Session
 
-from app.sqlmodels import OrgGroup, Taxon, TenkmRule
+from app.sqlmodels import OrgGroup, Taxon, TenkmRule, DifficultyCode, DifficultyRule
 from app.verify.verify_models import VerifyPack, OrgGroupRules
 
 from ..mocks import mock_make_search_request
@@ -51,12 +51,11 @@ class TestVerify:
             assert verified['name'] == "Adalia bipunctata"
             assert verified['date'] == "03/04/2024"
             assert verified['sref']['gridref'] == "TL123456"
-            assert verified['ok'] is False
-            assert len(verified['messages']) == 2
+            assert verified['result'] == 'warn'
+            assert 'id_difficulty' not in verified
+            assert len(verified['messages']) == 1
             assert verified['messages'][0] == (
-                "*:*:phenology: There is no rule for this taxon.")
-            assert verified['messages'][1] == (
-                "*:*:tenkm: There is no rule for this taxon.")
+                "*:*: No rules exist for this taxon.")
 
             # Update to test against specific org_group.
             rules = OrgGroupRules(
@@ -70,6 +69,8 @@ class TestVerify:
             )
             assert response.status_code == 200
             verified = response.json()['records'][0]
+            assert verified['result'] == 'fail'
+            assert 'id_difficulty' not in verified
             assert len(verified['messages']) == 1
             assert verified['messages'][0] == (
                 "Unrecognised organisation:group, 'UK Ladybird Survey:UKLS'.")
@@ -79,6 +80,7 @@ class TestVerify:
                 organisation='UK Ladybird Survey', group='UKLS')
             db.add(org_group)
             db.commit()
+
             # Now try again with that test.
             response = client.post(
                 '/verify',
@@ -86,11 +88,59 @@ class TestVerify:
             )
             assert response.status_code == 200
             verified = response.json()['records'][0]
+            assert verified['result'] == 'warn'
+            assert 'id_difficulty' not in verified
+            assert len(verified['messages']) == 1
+            assert verified['messages'][0] == (
+                "UK Ladybird Survey:UKLS: No rules exist for this taxon.")
+
+            # Create the missing difficulty rule.
+            # Create taxa.
+            taxon = Taxon(
+                name='Adalia bipunctata',
+                preferred_name='Adalia bipunctata',
+                search_name='adaliabipunctata',
+                tvk='NBNSYS0000008319',
+                preferred_tvk='NBNSYS0000008319',
+                preferred=True
+            )
+            db.add(taxon)
+            db.commit()
+            db.refresh(taxon)
+
+            # Create difficulty code.
+            difficulty_code = DifficultyCode(
+                code=1,
+                text='Easy',
+                org_group_id=org_group.id
+            )
+            db.add(difficulty_code)
+            db.commit()
+            db.refresh(difficulty_code)
+
+            # Create difficulty rule.
+            difficulty_rule = DifficultyRule(
+                org_group_id=org_group.id,
+                taxon_id=taxon.id,
+                difficulty_code_id=difficulty_code.id
+            )
+            db.add(difficulty_rule)
+            db.commit()
+
+            # Now try again with that test.
+            response = client.post(
+                '/verify',
+                json=pack.model_dump(),
+            )
+            assert response.status_code == 200
+            verified = response.json()['records'][0]
+            assert verified['result'] == 'warn'
+            assert verified['id_difficulty'] == 1
             assert len(verified['messages']) == 2
             assert verified['messages'][0] == (
-                "UK Ladybird Survey:UKLS:phenology: There is no rule for this taxon.")
+                "No rules run.")
             assert verified['messages'][1] == (
-                "UK Ladybird Survey:UKLS:tenkm: There is no rule for this taxon.")
+                "UK Ladybird Survey:UKLS:difficulty:1: Easy")
 
             # Update to test against specific rule.
             pack.org_group_rules_list[0].rules = ['tenkm']
@@ -101,27 +151,19 @@ class TestVerify:
             )
             assert response.status_code == 200
             verified = response.json()['records'][0]
-            assert len(verified['messages']) == 1
+            assert verified['result'] == 'warn'
+            assert verified['id_difficulty'] == 1
+            assert len(verified['messages']) == 2
             assert verified['messages'][0] == (
-                "UK Ladybird Survey:UKLS:tenkm: There is no rule for this taxon.")
+                "No rules run.")
+            assert verified['messages'][1] == (
+                "UK Ladybird Survey:UKLS:difficulty:1: Easy")
 
-            # Create a tenkm rule to test against.
-            # Create taxa.
-            taxon1 = Taxon(
-                name='Adalia bipunctata',
-                preferred_name='Adalia bipunctata',
-                search_name='adaliabipunctata',
-                tvk='NBNSYS0000008319',
-                preferred_tvk='NBNSYS0000008319',
-                preferred=True
-            )
-            db.add(taxon1)
-            db.commit()
-
-            # Create tenkm rule for org_group and taxon1.
+            # Create a tenkm to test against.
+            # Create tenkm rule for org_group and taxon.
             rule1 = TenkmRule(
                 org_group_id=org_group.id,
-                taxon_id=taxon1.id,
+                taxon_id=taxon.id,
                 km100='TL',
                 km10='14 15 16',
                 coord_system='OSGB'
@@ -136,8 +178,13 @@ class TestVerify:
             )
             assert response.status_code == 200
             verified = response.json()['records'][0]
-            assert len(verified['messages']) == 0
-            assert verified['ok'] is True
+            assert verified['result'] == 'pass'
+            assert verified['id_difficulty'] == 1
+            assert len(verified['messages']) == 2
+            assert verified['messages'][0] == (
+                'Rules run: tenkm')
+            assert verified['messages'][1] == (
+                'UK Ladybird Survey:UKLS:difficulty:1: Easy')
 
             # Change record location to be outside the tenkm rule - should fail.
             pack.records[0].sref.gridref = "TL 654 321"
@@ -147,10 +194,13 @@ class TestVerify:
             )
             assert response.status_code == 200
             verified = response.json()['records'][0]
-            assert len(verified['messages']) == 1
+            assert verified['result'] == 'fail'
+            assert verified['id_difficulty'] == 1
+            assert len(verified['messages']) == 2
             assert verified['messages'][0] == (
+                'UK Ladybird Survey:UKLS:difficulty:1: Easy')
+            assert verified['messages'][1] == (
                 "UK Ladybird Survey:UKLS:tenkm: Record is outside known area.")
-            assert verified['ok'] is False
 
             # Remove the rule list - should still fail.
             pack.org_group_rules_list = []
@@ -160,9 +210,10 @@ class TestVerify:
             )
             assert response.status_code == 200
             verified = response.json()['records'][0]
+            assert verified['result'] == 'fail'
+            assert verified['id_difficulty'] == 1
             assert len(verified['messages']) == 2
             assert verified['messages'][0] == (
-                "*:*:phenology: There is no rule for this taxon.")
+                'UK Ladybird Survey:UKLS:difficulty:1: Easy')
             assert verified['messages'][1] == (
                 "UK Ladybird Survey:UKLS:tenkm: Record is outside known area.")
-            assert verified['ok'] is False